Get PTSample loop state

Determines whether a loop is specified for a PTSample object.

Usage

# S4 method for class 'PTSample'
loopState(sample)

Arguments

sample

A PTSample object for which the loop state needs to be determined.

Value

Returns a logical value indicating whether a loop is (TRUE) or isn't (FALSE) specified for the sample.

Details

The loop state is not explicitly stored in a PTSample object. It can be derived from the loopStart position and loopLength. This method is provided as a convenient method to get the state. Use either loopStart or loopLength to change the state.

See also

Author

Pepijn de Vries

Examples

data("mod.intro")

## Get the loop status of sample number 1
## (it has a loop):
loopState(PTSample(mod.intro, 1))
#> [1] TRUE

## Get the loop status of sample number 2
## (it has no loop):
loopState(PTSample(mod.intro, 2))
#> [1] FALSE