Example of a PTModule object

A PTModule object included in the package as example.

Format

A PTModule object containing 4 PTSample objects (and 27 empty PTSample objects, adding up to the 31 samples a PTModule should hold) and 4 PTPattern objects.

Details

This PTModule object is based on an original ProTracker module file I've composed in the late nineteen nineties. It is used as example for many of the ProTrackR methods and you can use it to test your own code. It can also be exported back to the original ProTracker module file by using write.module.

Author

Pepijn de Vries

Examples

data("mod.intro")
print(mod.intro)
#> 
#> PTModule Object:
#> 	Module name:			intro			
#> 	Number of samples:		4		
#> 	Number of patterns:		4		
#> 	Pattern order table length:	9	
plot(mod.intro)
#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

if (FALSE) { # \dontrun{
playSample(mod.intro)

## Save as an original module file,
## which can be played with ProTracker (or several modern audio players):
write.module(mod.intro, "intro.mod")
} # }