Play PTModule objects

Converts PTModule objects into audio tuneR::Waves, and plays them.

Usage

# S4 method for class 'PTModule'
playMod(mod, wait = TRUE, ...)

Arguments

mod

A PTModule object to be played.

wait

A logical value. When set to TRUE the playing routine will wait with executing any code until the playing is finished. When set to FALSE, subsequent R code will be executed while playing.

...

Arguments that are passed on to modToWave.

Value

A tuneR::Wave object, generated from the mod object, is returned.

Details

Unfortunately, it was not feasible to create a routine that can directly interpret PTModule objects and play them simultaneously. Instead, the audio first needs to be rendered after which it can be played. This method therefore first calls modToWave and then playWave. Rendering may take some time and requires some balance between speed, quality and accuracy. See the documentation of the modToWave method for the control you have on these aspects.

See also

Author

Pepijn de Vries

Examples

if (interactive()) {
  data("mod.intro")

  ## play the module and capture the audio Wave
  wav <- playMod(mod.intro)
}