Play Wave objects

Use the command line audio::play function from the audio package to play tuneR::Wave objects.

Usage

# S4 method for class 'Wave'
playWave(wave, wait = TRUE)

# S4 method for class 'WaveMC'
playWave(wave, wait = TRUE)

Arguments

wave

An object of class tuneR::Wave or tuneR::WaveMC. Note that the playing routine implemented here can only play stereo waves. Multi-channel waves are therefore converted to stereo before playing.

wait

A logical value. When set to TRUE the playing routine will wait with executing any code until the playing is finished. When set to FALSE, subsequent R code will be executed while playing.

Value

Returns an audio::$.audioInstance.

Details

As the tuneR package play-function relies on external players, this method is provided as a convenient approach to play samples in the R console, using the audio package. Wave objects are played at the rate as specified in the object. Of course you can also play the Wave objects with the tuneR implementation of tuneR::play, by calling tuneR::play(wave).

See also

Author

Pepijn de Vries

Examples

if (interactive()) {
  data(mod.intro)

  ## PTSample objects can also be
  ## played with this function as they
  ## are a child of the Wave object:
  playWave(PTSample(mod.intro, 2))
}