Convert raw vectors into a character string

A function that converts raw data into a character string.

Usage

rawToCharNull(raw_dat)

Arguments

raw_dat

A vector of class raw to be converted into a character.

Value

A character string based on the raw data

Details

The function rawToChar() will fail on vectors of raw data with embedded 0x00 data. This function will not fail on embedded 0x00 values. Instead, it will replace embedded 0x00 data with white spaces. Note that leading and trailing 0x00 data will be omitted from the result.

See also

Author

Pepijn de Vries

Examples

## generate some raw data with an embedded 0x00:
some.raw.data <- as.raw(c(0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x00,
                          0x77, 0x6f, 0x72, 0x6c, 0x64, 0x21))
if (FALSE) { # \dontrun{
## this will fail:
try(rawToChar(some.raw.data))
} # }

## this will succeed:
rawToCharNull(some.raw.data)
#> [1] "hello world!"