Convert raw vector into a single unsigned integer value

This function converts raw data into an unsigned integer

Usage

rawToUnsignedInt(raw_dat)

Arguments

raw_dat

A vector of class raw to be converted into an unsigned integer

Value

A single unsigned integer value based on the provided raw data

Details

This function converts a vector of raw data into a single unsigned integer. for conversion of raw data into a vector of unsigned integers \[0,255\] use as.integer(). For an inverse of this function see unsignedIntToRaw().

See also

Author

Pepijn de Vries

Examples

## generate some raw data:
some.raw.data <- as.raw(c(0x01, 0x1e, 0x3f))

## convert raw data into an unsigned integer:
rawToUnsignedInt(some.raw.data)
#> [1] 73279

## note the difference with
as.integer(some.raw.data)
#> [1]  1 30 63