Get the length of a PTSample

Gets the length (in samples = bytes) of an audio fragment stored as a PTSample.

Usage

# S4 method for PTSample
sampleLength(sample)

Arguments

sample

A PTSample object for which the length needs to be returned.

Value

Returns a numeric value representing the number of samples (bytes) the PTSample object sample is composed of.

Details

PTSamples are 8 bit mono audio fragments. This method returns the length of this fragment expressed as number of samples (which also equals the number of bytes).

See also

Author

Pepijn de Vries

Examples

data("mod.intro")

## Show the length of the second sample in mod.intro
sampleLength(PTSample(mod.intro, 2))
#> [1] 1930